+++ tango/integrations/__init__.py
@@ -1,0 +1,4 @@
+"""
+In :mod:`tango.integrations` we provide many ready-to-use `component <../components/index.html>`_
+implementations for leveraging the functionality from popular libraries.
+"""