+++ CHANGELOG.md
@@ -6,3 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+
+- Ported over tango components from AllenNLP.
+
+## [v0.0.1](https://github.com/allenai/tango/releases/tag/v0.0.1) - 2021-09-22
+
+### Added
+
+- Added initial project boilerplate.